Ctrl+C also removes icon from sys tray, inc version number, removed qui entry  #65
