+++ keycat/__init__.py
@@ -1,1 +1,1 @@
-_VERSION = '0.1.12'
+_VERSION = '0.1.13'
